refactor:  Rename `synthetic_data.py`
