+++ examples/helpers/synthetic_data.py
